Move parser config to new function
