+++ run_seas5.py
@@ -1,4 +1,3 @@
-import argparse
 import logging
 import tempfile
 from datetime import datetime
@@ -7,63 +6,16 @@ from pathlib import Path
 import src.seas5.aws_tprate as aws_tprate
 import src.seas5.mars_tprate as mars_tprate
 from constants import BBOX_GLOBAL, BBOX_TEST
+from src.seas5.set_inputs import cli_args
 
 logger = logging.getLogger(__name__)
-
-
-def check_range(value):
-    ivalue = int(value)
-    if ivalue < 1981 or ivalue > 2022:
-        raise argparse.ArgumentTypeError(
-            f"Value {value} is outside the valid range (1981-2022)"
-        )
-    return ivalue
 
 
 if __name__ == "__main__":
     logging.basicConfig()
     logging.getLogger().setLevel(logging.INFO)
 
-    parser = argparse.ArgumentParser()
-    parser.add_argument(
-        "--start",
-        "-s",
-        help="""Start year to retrieve and process archival SEAS5 data.
-        Must be between 1981 and 2022 (default: 1981). Only applies for `--source mars`""",
-        default=1981,
-        type=check_range,
-    )
-    parser.add_argument(
-        "--end",
-        "-e",
-        help="""End year to retrieve and process archival SEAS5 data.
-        Must be between 1981 and 2022 (default: 2022). Only applies for `--source mars`""",
-        default=2022,
-        type=check_range,
-    )
-    parser.add_argument(
-        "--mode",
-        "-m",
-        help="Run the pipeline in 'local', 'dev', or 'prod' mode.",
-        type=str,
-        choices=["local", "dev", "prod"],
-        default="local",
-    )
-    parser.add_argument(
-        "--source",
-        "-src",
-        help="Data download source",
-        type=str,
-        choices=["mars", "aws"],
-    )
-    parser.add_argument(
-        "--backfill-aws",
-        help="""Will backfill all previous months of AWS data.
-        If not flagged, only data from the current month will be processed.""",
-        action="store_true",
-    )
-
-    args = parser.parse_args()
+    args = cli_args()
 
     if args.source == "mars":
         logger.info(f"Retrieving SEAS5 archive from {args.start} to {args.end}...")
