+++ src/seas5/README.md
@@ -1,52 +1,61 @@
 ## Usage
 
-The pipeline can be run locally from the command line:
+The pipeline can be run from the command line as follows:
 
 ```
-usage: run_seas5.py [-h] [--start START] [--end END] [--test]
+usage: run_seas5.py [-h] [--start START] [--end END] [--mode {local,dev,prod}] [--source {mars,aws}] [--backfill-aws]
 
 options:
   -h, --help            show this help message and exit
   --start START, -s START
-                        Start year to retrieve and process archival SEAS5 data. Must be
-                        between 1981 and 2022 (default: 1981).
-  --end END, -e END     End year to retrieve and process archival SEAS5 data. Must be
-                        between 1981 and 2022 (default: 2022).
-  --test, -t            Run the pipeline in test mode. Will save a subset of outputs
-                        locally to 'test_outputs/' and not upload any data to Azure.
+                        Start year to retrieve and process archival SEAS5 data. Must be between 1981 and 2022 (default:
+                        1981). Only applies for `--source mars`
+  --end END, -e END     End year to retrieve and process archival SEAS5 data. Must be between 1981 and 2022 (default:
+                        2022). Only applies for `--source mars`
+  --mode {local,dev,prod}, -m {local,dev,prod}
+                        Run the pipeline in 'local', 'dev', or 'prod' mode.
+  --source {mars,aws}, -src {mars,aws}
+                        Data download source
+  --backfill-aws        Will backfill all previous months of AWS data. If not flagged, only data from the current month
+                        will be processed.
 ```
-
-When not run in test mode, this will create outputs in two places:
-
-1) A single raw `.grib` file for each year will be saved to the `dev` Azure storage container under `global/mars/raw/`
-2) For each year, 84 `.tif` files will be saved to the `prod` Azure storage container under `raster/seas5/`. See the section below for more details.
 
 This code is also configured as a Job on Databricks, called "Update SEAS5 Archive". This can be triggered manually and has been used for bulk tasks (ie. more than a couple years) due to significantly improved performance.
 
 ### Example usage
 
-1. Process the full MARS archive from 1981 to 2022 and save all outputs to appropriate locations on Azure:
+1. Process the full MARS archive from 1981 to 2022 and save all outputs to `prod` Azure container:
 
 ```
-python run_seas5.py
+python run_seas5.py -src mars -m prod
 ```
 
-2. Process data from 2000 to 2010 and save all outputs to appropriate locations on Azure:
+2. Process MARS data from 2000 to 2010 and save all outputs to `dev` Azure container
 
 ```
-python run_seas5.py -s 2000 -e 2011
+python run_seas5.py -src mars -s 2000 -e 2011 -m dev
 ```
 
-3. Test the pipeline by locally downloading and processing data only from 1990:
+3. Process the current month's outputs from AWS and save locally
 
 ```
-python run_seas5.py -s 1990 -e 1990 -t
+python run_seas5.py -src aws
+```
+
+3. Backfill all data from AWS bucket and save to `prod` Azure container
+
+```
+python run_seas5.py -src aws --backfill-aws -m prod
 ```
 
 ## Processing details
 
-### Raw files:
-Global, monthly precipitation forecasts are downloaded in yearly `.grib` files. Each raw `.grib` contains all ensemble members (26 or 51, depending on the year) and lead times (0-6 months ahead). See [this JIRA ticket](https://humanitarian.atlassian.net/browse/DSCI-539?focusedCommentId=177527) for more detailed docs on how the MARS API call is parameterized. All raw `.grib` data is stored in the `dev` Azure storage container under `global/mars/raw/`. Files are named `seas5_mars_tprate_{year}.grib`.
+### Raw MARS files:
+Global, monthly precipitation forecasts are downloaded in yearly `.grib` files. Each raw `.grib` contains all ensemble members (26 or 51, depending on the year) and lead times (0-6 months ahead). See [this JIRA ticket](https://humanitarian.atlassian.net/browse/DSCI-539?focusedCommentId=177527) for more detailed docs on how the MARS API call is parameterized. All raw `.grib` data is stored in Azure storage container under `raster/seas5/mars/raw/`.
+
+### Raw AWS files:
+
+TODO
 
 ### Processed files:
 The `.grib` file from each year is processed to output 84 cloud-optimized-geotiffs (`.tif`):
@@ -54,4 +63,4 @@ The `.grib` file from each year is processed to output 84 cloud-optimized-geotif
 2. Separate by publication month and lead time
 3. Set a CRS (`EPSG:4326`)
 
-All processed files are saved to the `prod` Azure storage container under `raster/seas5/`. Files are named `seas5_mars_tprate_em_i{pub_date}_lt{leadtime}.tif`. Note that `lt0` is when valid_date=pub_date.
+All processed files are saved to the `prod` Azure storage container under `raster/seas5/*/processed/`. Files are named `tprate_em_i{pub_date}_lt{leadtime}.tif`. Note that `lt0` is when valid_date=pub_date.
